feat(toc): display active section in article toc
